adding _animateHide attempt + restructure
