+++ mixins/visibility/demo/visibility-mixin.html
@@ -19,8 +19,14 @@
 
 			<d2l-demo-snippet>
 				<template>
-					<d2l-visibility-test-helper></d2l-visibility-test-helper>
-					<input type='checkbox'>Temp</button>
+					<h5>Group 1</h5>
+					<div>
+						<d2l-visibility-test-helper></d2l-visibility-test-helper>
+					</div>
+					<h5>Group 2</h5>
+					<div>
+						<input type='checkbox'>Temp</button>
+					</div>
 				</template>
 			</d2l-demo-snippet>
