+++ mixins/visibility/demo/visibility-test-helper.js
