+++ mixins/visibility/demo/visibility-test.js
